+++ CHANGELOG.md
@@ -7,6 +7,8 @@ of how the versioning scheme that this project does not use is clearly superior
 to random natural numbers and how incrementing the minor version number
 indicates backward compatible changes, unless it does not.
 
+- 2019-09-29 — Removed `make-widget`. Widget creation is now done with
+  Common Lisp's `make-instance`.
 - 2019-09-29 — Add `combobox`, `file-upload` and support for v7.5 widgets.
 - 2019-09-27 — Add support for JupyterLab in Docker.
 - 2019-09-07 — Add system-wide installation method.